Migrate httpdportalpreview to caddy2
