+++ go/cmd/pfhttpd2/main.go
@@ -8,6 +8,7 @@ import (
 	_ "github.com/inverse-inc/packetfence/go/plugin/caddy2/api"
 	_ "github.com/inverse-inc/packetfence/go/plugin/caddy2/api-aaa"
 	_ "github.com/inverse-inc/packetfence/go/plugin/caddy2/httpdispatcher"
+	_ "github.com/inverse-inc/packetfence/go/plugin/caddy2/httpdportalpreview"
 	_ "github.com/inverse-inc/packetfence/go/plugin/caddy2/pfconfig"
 )
